backend/requirements.txt
Flask==2.2.2
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.2.0
scikit_image==0.19.3
skimage==0.0
tensorflow==2.10.0
